% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltTextSubset.R
\docType{data}
\name{AltTextSubset}
\alias{AltTextSubset}
\title{TidyTuesday tweets from 2018-2021 that contain media and customized alternative (alt) text}
\format{
A tibble with 465 rows and 7 variables:
\describe{
\item{TweetId}{<chr> Unique tweet identifier.}
\item{ImageUrl}{<chr> URL to the media shared in the tweet.}
\item{AltText}{<chr> Alternative text corresponding to the media shared in the tweet.

There are three types of entries:
\enumerate{
\item Customized alternative text written by the tweet author.
\item "ScrapingError", denoting unavailable alternative text due to unavailable tweets, like private or deleted tweets.
\item \code{NA}, denoting unavailable alternative text due to the media being non-static, like a gif or a video.}
}

\item{HashtagList}{<list> List of hashtags used in the tweet.}
\item{TweetDate}{<dttm> Date and time the tweet was posted. Ranges from April 10, 2018 to April 4, 2021.}
\item{Year}{<fct> Year the tweet was posted. Includes 2018-2021.}
\item{UrlCheck}{<fct> Denotes whether the tweet included an external link.

There are two types of entries:
\itemize{
\item "IncludesLinks" which denotes the tweet included external links in addition to media, like a link to source code.
\item "NoLinks" which denotes the tweet did not include external links (i.e. only included media).}
}
}
}
\source{
{Tweets originally collected by Thomas Mock. Thomas Mock (2021). Tidy Tuesday: A weekly data project aimed at the R ecosystem.} \url{https://github.com/rfordatascience/tidytuesday}

{Alternative text for tweets was scraped by Silvia Canelón after further processing.} \url{https://www.twitter.com}
}
\usage{
AltTextSubset
}
\description{
A dataset containing the alternative text for media shared between 2018 and 2021 as part of the TidyTuesday social project, along with other tweet attributes.
This is a subset of the 2018-2021 datasets, containing only tweets with alternative text that isn't "Image," the default alternative text added by the Twitter app in the absence of customized alternative text.
}
\keyword{datasets}
