% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyTuesdayAltText-package.R
\docType{package}
\name{TidyTuesdayAltText-package}
\alias{TidyTuesdayAltText}
\alias{TidyTuesdayAltText-package}
\title{TidyTuesdayAltText: Alternative text for media attached to TidyTuesday tweets}
\description{
Alternative (alt) text for media attached to tweets participating in the TidyTuesday social project (i.e. using the #TidyTuesday hashtag). Most often the media is an image of a data visualization.
}
\author{
\strong{Maintainer}: Silvia Canelón \email{sp.canelon@gmail.com} (\href{https://orcid.org/0000-0003-1709-1394}{ORCID})

Authors:
\itemize{
  \item Thomas Mock \email{j.thomasmock@gmail.com}
}

}
\keyword{internal}
